Fix reviewer suggestions
