--- src/components/Books.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Books = () => (
-  <h1>Books Page</h1>
-);
-
-export default Books;
