--- src/components/Categories.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Categories = () => (
-  <h1>Under Construction</h1>
-);
-
-export default Categories;
