+++ src/redux/books/Books.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Books = () => (
+
+  <form>
+    <input type="text" placeholder="Book Title" />
+    <input type="text" placeholder="Book Author" />
+    <button type="button">Add Book</button>
+  </form>
+
+);
+
+export default Books;
