+++ src/redux/categories/Categories.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Categories = () => (
+  <h1>Under Construction</h1>
+);
+
+export default Categories;
